Scrape trophies script
Quick script to take screen shots of undefeated trophy counts. Must
start parallels in “more real estate” mode, maximize modo on the right
side of the screen and drag its left side all the way to the right. The
movement in line 5 assumes that the parallels windows thumbnail will be
at that position.